+++ src/main/java/com/robertx22/age_of_exile/api/MineAndSlashEvents.java
@@ -1,10 +1,10 @@
 package com.robertx22.age_of_exile.api;
 
 import com.robertx22.age_of_exile.capability.entity.EntityCap;
+import com.robertx22.age_of_exile.damage_hooks.util.AttackInformation;
 import com.robertx22.age_of_exile.event_hooks.my_events.CollectGearEvent;
 import com.robertx22.age_of_exile.saveclasses.unit.GearData;
 import com.robertx22.age_of_exile.uncommon.datasaving.Load;
-import com.robertx22.age_of_exile.uncommon.effectdatas.AttackInformation;
 import net.minecraft.entity.LivingEntity;
 
 import java.util.List;