+++ src/main/java/com/robertx22/age_of_exile/damage_hooks/OnNonPlayerDamageEntityEvent.java
@@ -1,8 +1,8 @@
-package com.robertx22.age_of_exile.event_hooks.my_events;
+package com.robertx22.age_of_exile.damage_hooks;
 
+import com.robertx22.age_of_exile.damage_hooks.util.AttackInformation;
 import com.robertx22.age_of_exile.database.data.spells.spell_classes.bases.MyDamageSource;
 import com.robertx22.age_of_exile.mixin_methods.OnHurtEvent;
-import com.robertx22.age_of_exile.uncommon.effectdatas.AttackInformation;
 import com.robertx22.library_of_exile.events.base.EventConsumer;
 import com.robertx22.library_of_exile.events.base.ExileEvents;
 import net.minecraft.entity.player.PlayerEntity;
@@ -15,6 +15,9 @@ public class OnNonPlayerDamageEntityEvent extends EventConsumer<ExileEvents.OnDa
             return;
         }
         if (event.source instanceof MyDamageSource) {
+            return;
+        }
+        if (LivingHurtUtils.isEnviromentalDmg(event.source)) {
             return;
         }
         if (!(event.source.getAttacker() instanceof PlayerEntity)) {