+++ src/main/java/com/robertx22/age_of_exile/damage_hooks/util/DmgSourceUtils.java
@@ -1,4 +1,4 @@
-package com.robertx22.age_of_exile.event_hooks.entity.damage;
+package com.robertx22.age_of_exile.damage_hooks.util;
 
 import com.robertx22.age_of_exile.database.data.spells.spell_classes.bases.MyDamageSource;
 import com.robertx22.age_of_exile.uncommon.effectdatas.DamageEffect;