+++ src/main/java/com/robertx22/age_of_exile/database/data/spells/components/actions/DamageAction.java
@@ -59,7 +59,6 @@ public class DamageAction extends SpellAction implements ICTextTooltip {
                     dmg.effectType = data.getDmgEffectType();
                 }
                 dmg.element = ele;
-
                 dmg.Activate();
             }
         }