+++ CHANGELOG.md
@@ -75,4 +75,5 @@ Now you can enable compat only for certain mods.
 * added death statistics overlay when dead. 
 Now you can see what damage types you died to. 
 So if you die to mostly fire damage for example, you should know to up your resist.
+* added hurt sounds to spells
 * added chat warning if you have immersive portals installed (configurable)